store/tikv: move mocktikv into tikv
